micro-service-geo/requirements.txt
aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
elevation==1.1.3
fastapi==0.109.2
fasteners==0.19
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
idna==3.6
multidict==6.0.5
numpy==1.26.4
orjson==3.9.13
polyline==2.0.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.7
PyYAML==6.0.1
shapely==2.0.2
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
